[#19] refactor: common Level 타입 제거
